+++ emails-please/src/App.css
@@ -1,158 +1,149 @@
 .app-container {
-    display: grid;
-    grid-template-columns: 175px 1fr;
-    height: 100vh;
+  display: grid;
+  grid-template-columns: 175px 1fr;
+  height: 100vh;
 }
-
 
 .content-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 100%;
-    max-width: 900px;
-    margin: 0 auto;
-    padding: 20px;
-    box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  max-width: 900px;
+  margin: 0 auto;
+  padding: 20px;
+  box-sizing: border-box;
 }
-
 
 .email .email-from,
 .email .email-to {
-    display: inline-flex;
-    align-items: center;
-    box-sizing: border-box;
+  display: inline-flex;
+  align-items: center;
+  box-sizing: border-box;
 }
-
 
 .email-container {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: stretch;
-    padding: 20px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: stretch;
+  padding: 20px;
 }
-
 
 .popup {
-    position: fixed;
-    top: 20px;
-    left: 50%;
-    transform: translateX(-50%);
-    padding: 15px 30px;
-    border-radius: 8px;
-    font-size: 1.2rem;
-    z-index: 1000;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.2);
-    animation: fade-in-out 5s ease;
+  position: fixed;
+  top: 20px;
+  left: 50%;
+  transform: translateX(-50%);
+  padding: 15px 30px;
+  border-radius: 8px;
+  font-size: 1.2rem;
+  z-index: 1000;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.2);
+  animation: fade-in-out 5s ease;
 }
-
 
 .popup.success {
-    background-color: #4caf50;
-    color: white;
+  background-color: #4caf50;
+  color: white;
 }
-
 
 .popup.error {
-    background-color: #f44336;
-    color: white;
+  background-color: #f44336;
+  color: white;
 }
 
-
 @keyframes fade-in-out {
-    0% {
-        opacity: 0;
-        transform: translateX(-50%) translateY(-20px);
-    }
-    10% {
-        opacity: 1;
-        transform: translateX(-50%) translateY(0);
-    }
-    90% {
-        opacity: 1;
-        transform: translateX(-50%) translateY(0);
-    }
-    100% {
-        opacity: 0;
-        transform: translateX(-50%) translateY(-20px);
-    }
+  0% {
+    opacity: 0;
+    transform: translateX(-50%) translateY(-20px);
+  }
+  10% {
+    opacity: 1;
+    transform: translateX(-50%) translateY(0);
+  }
+  90% {
+    opacity: 1;
+    transform: translateX(-50%) translateY(0);
+  }
+  100% {
+    opacity: 0;
+    transform: translateX(-50%) translateY(-20px);
+  }
 }
 
 .controls {
-    align-self: center;
-    margin-top: 20px;
+  align-self: center;
+  margin-top: 20px;
 }
 
 .sidebar {
-    background-color: #f4f4f4;
-    border-right: 1px solid #ccc;
-
+  background-color: #f4f4f4;
+  border-right: 1px solid #ccc;
 }
 
 .email-viewer {
-    border: 1px solid #ccc;
-    padding: 20px;
-    margin-bottom: 20px;
+  border: 1px solid #ccc;
+  padding: 20px;
+  margin-bottom: 20px;
 }
 
 .controls {
-    display: flex;
-    justify-content: flex-end;
-    gap: 10px;
+  display: flex;
+  justify-content: flex-end;
+  gap: 10px;
 }
 
 button {
-    width: 50px;
-    height: 50px;
-    border: none;
-    cursor: pointer;
+  width: 50px;
+  height: 50px;
+  border: none;
+  cursor: pointer;
 }
 
 button.red {
-    background-color: red;
+  background-color: red;
 }
 
 button.green {
-    background-color: green;
+  background-color: green;
 }
 
-
 .points-counter {
-    position: fixed;
-    top: 20px;
-    right: 20px;
-    background-color: #ffffff;
-    color: black;
-    padding: 10px 20px;
-    border-radius: 8px;
-    font-size: 1.2rem;
-    z-index: 1000;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.2);
+  position: fixed;
+  top: 20px;
+  right: 20px;
+  background-color: #ffffff;
+  color: black;
+  padding: 10px 20px;
+  border-radius: 8px;
+  font-size: 1.2rem;
+  z-index: 1000;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.2);
 }
 
 .controls {
-    display: flex;
-    justify-content: center;
-    gap: 20px;
-    margin-top: 20px;
+  display: flex;
+  justify-content: center;
+  gap: 20px;
+  margin-top: 20px;
 }
 
 button {
-    padding: 10px 20px;
-    font-size: 1rem;
-    border: 2px solid #000;
-    border-radius: 8px;
-    background-color: #fff;
-    cursor: pointer;
-    transition: background-color 0.3s ease, transform 0.2s ease;
+  padding: 10px 20px;
+  font-size: 1rem;
+  border: 2px solid #000;
+  border-radius: 8px;
+  background-color: #fff;
+  cursor: pointer;
+  transition: background-color 0.3s ease, transform 0.2s ease;
 }
 
 button:hover {
-    background-color: #f0f0f0;
+  background-color: #f0f0f0;
 }
 
 button:disabled {
-    background-color: #ccc;
-    cursor: not-allowed;
+  background-color: #ccc;
+  cursor: not-allowed;
 }
